+++ Assets/Scripts/TargetScripts/Target.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 public class Target : MonoBehaviour, IDamageable
 {
     public event Action OnGetDamaged;
-    public void GetDamage(int damageAmount)
+    public void GetDamage(float damageAmount)
     {
         OnGetDamaged?.Invoke();
     }
